(feature, ci) Does vertex radius match timevalue?
Add a function to check the squared distance of a vertex from the origin and compare it to the squared timevalue. These should match for proper vertices. This required adding new data members to remember the initial radius and spacing of the foliations in the Foliated Triangulation.

Because we are comparing squared distances using doubles, add a TOLERANCE interval for the comparison.

Also change Travis-CI to require TSAN to pass.

425/431 tests passing.

+++ src/initialize.cpp
@@ -1,6 +1,6 @@
 /// Causal Dynamical Triangulations in C++ using CGAL
 ///
-/// Copyright © 2014-2020 Adam Getchell
+/// Copyright © 2014-2021 Adam Getchell
 ///
 /// A program that generates spacetimes
 
@@ -54,8 +54,8 @@ try
   auto simplices         = stoll(args["-n"].asString());
   auto timeslices        = stoll(args["-t"].asString());
   auto dimensions        = stoll(args["-d"].asString());
-  auto initial_radius    = stold(args["--init"].asString());
-  auto foliation_spacing = stold(args["--foliate"].asString());
+  auto initial_radius    = stod(args["--init"].asString());
+  auto foliation_spacing = stod(args["--foliate"].asString());
   auto save_file         = args["--output"].asBool();
 
   // Initialize triangulation
